Move meta/assessment/assessmentFile -> meta/cycleData
