+++ src/client/store/ui/assessmentFiles/actions/getFiles.ts
@@ -3,7 +3,7 @@ import axios from 'axios'
 
 import { ApiEndPoint } from 'meta/api/endpoint'
 import { CycleParams } from 'meta/api/request'
-import { AssessmentFile } from 'meta/assessment'
+import { AssessmentFile } from 'meta/cycleData'
 
 export const getFiles = createAsyncThunk<Array<AssessmentFile>, CycleParams>('assessment/files/get', async (params) => {
   const { data } = await axios.get(ApiEndPoint.File.Assessment.many(), { params })
