+++ src/client/store/ui/assessmentFiles/stateType.ts
@@ -1,5 +1,5 @@
 import { CountryIso } from 'meta/area'
-import { AssessmentFile } from 'meta/assessment'
+import { AssessmentFile } from 'meta/cycleData'
 
 export type AssessmentFilesState = {
   [key in CountryIso | string]: Array<AssessmentFile>
