+++ src/meta/assessment/index.ts
@@ -3,7 +3,6 @@ export { ActivityLogMessage } from './activityLog'
 export { ActivityLogs } from './activityLogs'
 export type { Assessment } from './assessment'
 export type { AssessmentProps } from './assessment'
-export type { AssessmentFile } from './assessmentFile'
 export type {
   AssessmentMetaCache,
   DependencyCache,
@@ -67,3 +66,4 @@ export type { Table, TableColumnNames, TableName, TableProps } from './table'
 export { TableNames } from './table'
 export type { TableSection, TableSectionProps } from './tableSection'
 export { Unit } from './unit'
+export type { AssessmentFile } from 'meta/cycleData/assessmentFile'
