+++ src/meta/cycleData/index.ts
@@ -1,0 +1,1 @@
+export type { AssessmentFile } from './assessmentFile'
