+++ src/server/repository/assessment/file/getOne.ts
@@ -1,4 +1,5 @@
-import { Assessment, AssessmentFile } from 'meta/assessment'
+import { Assessment } from 'meta/assessment'
+import { AssessmentFile } from 'meta/cycleData'
 
 import { BaseProtocol, DB, Schemas } from 'server/db'
