+++ src/server/repository/assessment/file/remove.ts
@@ -1,6 +1,7 @@
 import { Objects } from 'utils/objects'
 
-import { Assessment, AssessmentFile } from 'meta/assessment'
+import { Assessment } from 'meta/assessment'
+import { AssessmentFile } from 'meta/cycleData'
 
 import { BaseProtocol, DB, Schemas } from 'server/db'
